refactor(mobile): split the markdown editor document from its WebView presentation
Split the editor document into a body module and a standalone script module so a
host that is not this WebView can serve the same program, and move the message
handling into a controller hook behind a transport. MobileRichMarkdownEditor
stays the only presentation and supplies the injectJavaScript transport. A
snapshot of main's document pins the built HTML byte for byte.
