+++ mobile/src/components/mobile-rich-markdown-editor-document-body.ts
@@ -1,4 +1,5 @@
-export const MOBILE_RICH_MARKDOWN_EDITOR_BODY_PRIMARY = [
+// Head of the editor document through the editable surface; the script follows it.
+export const MOBILE_RICH_MARKDOWN_EDITOR_DOCUMENT_BODY = [
   ';',
   '      --font-mono: ui-monospace, SFMono-Regular, Menlo, Monaco, Consolas, "Liberation Mono", monospace;',
   '      --font-sans: Geist, -apple-system, BlinkMacSystemFont, "Segoe UI", sans-serif;',
@@ -181,99 +182,5 @@ export const MOBILE_RICH_MARKDOWN_EDITOR_BODY_PRIMARY = [
   '  </style>',
   '</head>',
   '<body>',
-  '  <main id="editor" contenteditable="true" data-placeholder="Start writing..."></main>',
-  '  <script>',
-  '    (function () {',
-  "      var editor = document.getElementById('editor');",
-  "      var lastMarkdown = '';",
-  '      var inputTimer = null;',
-  '      var documentGeneration = 0;',
-  '      var editable = true;',
-  '      var suppressInput = false;',
-  '',
-  '      function post(message) {',
-  '        window.ReactNativeWebView && window.ReactNativeWebView.postMessage(JSON.stringify(message));',
-  '      }',
-  '',
-  '      function decodeMarkdownEntities(value) {',
-  '        return String(value).replace(/&(#x[0-9a-f]+|#\\d+|amp|lt|gt|quot|apos);/gi, function (match, entity) {',
-  '          var lower = String(entity).toLowerCase();',
-  "          if (lower === 'amp') return '&';",
-  "          if (lower === 'lt') return '<';",
-  "          if (lower === 'gt') return '>';",
-  "          if (lower === 'quot') return '\"';",
-  "          if (lower === 'apos') return \"'\";",
-  "          if (lower.indexOf('#x') === 0) {",
-  '            var hex = Number.parseInt(lower.slice(2), 16);',
-  '            return Number.isFinite(hex) && hex >= 0 && hex <= 0x10ffff ? String.fromCodePoint(hex) : match;',
-  '          }',
-  "          if (lower.indexOf('#') === 0) {",
-  '            var code = Number.parseInt(lower.slice(1), 10);',
-  '            return Number.isFinite(code) && code >= 0 && code <= 0x10ffff ? String.fromCodePoint(code) : match;',
-  '          }',
-  '          return match;',
-  '        });',
-  '      }',
-  '',
-  '      function escapeHtml(value) {',
-  '        return decodeMarkdownEntities(value).replace(/[&<>"\']/g, function (char) {',
-  "          return ({ '&': '&amp;', '<': '&lt;', '>': '&gt;', '\"': '&quot;', \"'\": '&#39;' })[char];",
-  '        });',
-  '      }',
-  '',
-  '      function escapeAttr(value) {',
-  "        return escapeHtml(value).replace(/\\n/g, ' ');",
-  '      }',
-  '',
-  '      function isSafeUrl(value) {',
-  "        var trimmed = String(value || '').trim();",
-  '        return !/^javascript:/i.test(trimmed);',
-  '      }',
-  '',
-  '      function splitTableRow(line) {',
-  "        return line.trim().replace(/^\\|/, '').replace(/\\|$/, '').split('|').map(function (cell) {",
-  '          return cell.trim();',
-  '        });',
-  '      }',
-  '',
-  '      function isTableSeparator(line) {',
-  '        var cells = splitTableRow(line);',
-  '        return cells.length > 0 && cells.every(function (cell) {',
-  '          return /^:?-{3,}:?$/.test(cell);',
-  '        });',
-  '      }',
-  '',
-  '      function renderInline(text) {',
-  "        var output = '';",
-  '        var pattern = /(!\\[[^\\]]*\\]\\([^)]+\\)|`[^`]+`|~~[^~]+~~|\\*\\*[^*]+\\*\\*|__[^_]+__|\\*[^*\\n]+\\*|_[^_\\n]+_|\\[[^\\]]+\\]\\([^)]+\\)|https?:\\/\\/[^\\s<]+)/g;',
-  '        var lastIndex = 0;',
-  '        var match;',
-  '        while ((match = pattern.exec(text))) {',
-  '          output += escapeHtml(text.slice(lastIndex, match.index));',
-  '          var token = match[0];',
-  '          var image = token.match(/^!\\[([^\\]]*)\\]\\(([^)]+)\\)$/);',
-  '          var link = token.match(/^\\[([^\\]]+)\\]\\(([^)]+)\\)$/);',
-  '          if (image && isSafeUrl(image[2])) {',
-  "            output += '<img src=\"' + escapeAttr(image[2]) + '\" alt=\"' + escapeAttr(image[1] || '') + '\" />';",
-  '          } else if (link && isSafeUrl(link[2])) {',
-  "            output += '<a href=\"' + escapeAttr(link[2]) + '\">' + renderInline(link[1]) + '</a>';",
-  '          } else if (/^https?:\\/\\//i.test(token)) {',
-  "            output += '<a href=\"' + escapeAttr(token) + '\">' + escapeHtml(token) + '</a>';",
-  "          } else if (token.indexOf('`') === 0) {",
-  "            output += '<code>' + escapeHtml(token.slice(1, -1)) + '</code>';",
-  "          } else if (token.indexOf('~~') === 0) {",
-  "            output += '<s>' + renderInline(token.slice(2, -2)) + '</s>';",
-  "          } else if (token.indexOf('**') === 0 || token.indexOf('__') === 0) {",
-  "            output += '<strong>' + renderInline(token.slice(2, -2)) + '</strong>';",
-  '          } else {',
-  "            output += '<em>' + renderInline(token.slice(1, -1)) + '</em>';",
-  '          }',
-  '          lastIndex = pattern.lastIndex;',
-  '        }',
-  '        output += escapeHtml(text.slice(lastIndex));',
-  '        return output;',
-  '      }',
-  '',
-  '      function isBlockStart(line) {',
-  '        return /^(```|#{1,6}\\s+|>\\s?|\\s*(?:[-*+]|\\d+[.)])\\s+|\\s*(-{3,}|\\*{3,}|_{3,})\\s*$)/.test(line);'
+  '  <main id="editor" contenteditable="true" data-placeholder="Start writing..."></main>'
 ].join('\n')
