+++ mobile/src/components/mobile-rich-markdown-editor-script-string.ts
@@ -1,0 +1,3 @@
+export function escapeInjectedJavaScriptString(value: string): string {
+  return JSON.stringify(value).replace(/<\/script/gi, '<\\/script')
+}
